ADSR state machine and other songs of the 80s
This branch is just a proof-of-concept for using a state-machine to handle the ADSR ramp up-down.

This is so that the current volume is preserved from the end of one note into the beginning of the next, whereas previously a "pop" effect would be heard on note transitions.

This engine still cannot handle maximum volume- something is overflowing perhaps, and causing horrible clipping and distortion. It's truly an earsore to behold.

We're still missing low-pass and band-pass filters. Or maybe cut-off and resonance? Something like that.

This seems to sound more or less the same in SDL as it does on 32Blit, but SDL has a sample buffer which should ideally be the same length as an update cycle.

However I don't think the SDL audio fetch is in-sync with the engine, so it may be producing audio mid-way through a register change. Hmm!

Anyway this has been fun, challenging, frustrating, interesting, and ear-piercing.